Fix: Fix common dto and seperate Ordertype to common types/enum
