+++ src/common/dto/index.ts
@@ -1,1 +1,2 @@
-export * from './pagination.dto';
+export * from './request';
+export * from './response';
