+++ src/common/dto/request/index.ts
@@ -1,0 +1,1 @@
+export * from './paginated.dto';
