+++ src/common/dto/response/index.ts
@@ -1,0 +1,1 @@
+export * from './paginated.dto';
