+++ src/common/types/enum.ts
@@ -1,0 +1,4 @@
+export enum OrderType {
+  asc = 'asc',
+  desc = 'desc',
+}
