+++ src/modules/classification/classification.controller.ts
@@ -8,7 +8,7 @@ import {
   Query,
   UseGuards,
 } from '@nestjs/common';
-import { GetUser, PaginationQuery } from '@src/common';
+import { BasePaginationQuery, GetUser } from '@src/common';
 import { JwtGuard } from '../users/guards';
 import { ClassificationService } from './classification.service';
 import {
@@ -48,7 +48,7 @@ export class ClassificationController {
   @GetAIPostListDocs
   async getSuggestedPostList(
     @GetUser() userId: string,
-    @Query() pagingQuery: PaginationQuery,
+    @Query() pagingQuery: BasePaginationQuery,
   ) {
     const { count, classificationPostList } =
       await this.classificationService.getPostList(userId, pagingQuery);
@@ -66,7 +66,7 @@ export class ClassificationController {
   async getSuggestedPostListInFolder(
     @GetUser() userId: string,
     @Param('folderId') folderId: string,
-    @Query() pagingQuery: PaginationQuery,
+    @Query() pagingQuery: BasePaginationQuery,
   ) {
     const { count, classificationPostList } =
       await this.classificationService.getPostListInFolder(
