+++ src/modules/classification/response/ai-post-list.dto.ts
@@ -1,5 +1,5 @@
 import { ApiProperty } from '@nestjs/swagger';
-import { BasePaginationResponse } from '@src/common/dto/response/paginated.dto';
+import { BasePaginationResponse } from '@src/common';
 import { ClassificationPostList } from '../dto/classification.dto';
 
 export class AIPostListResponse extends BasePaginationResponse<ClassificationPostList> {
