+++ src/modules/folders/folders.controller.ts
@@ -9,7 +9,7 @@ import {
   Query,
   UseGuards,
 } from '@nestjs/common';
-import { GetUser, PaginationMetadata } from '@src/common';
+import { GetUser } from '@src/common';
 import { ClassificationService } from '@src/modules/classification/classification.service';
 import { GetPostQueryDto } from '../posts/dto/find-in-folder.dto';
 import { PostsService } from '../posts/posts.service';
@@ -90,13 +90,8 @@ export class FoldersController {
       query,
     );
 
-    const metadata = new PaginationMetadata(
-      query.page,
-      query.limit,
-      result.count,
-    );
     const posts = result.posts.map((post) => new PostResponse(post));
-    return new FolderPostResponse(metadata, posts);
+    return new FolderPostResponse(query.page, query.limit, result.count, posts);
   }
 
   @UpdateFolderDocs
