+++ src/modules/posts/dto/find-in-folder.dto.ts
@@ -1,9 +1,9 @@
 import { ApiProperty } from '@nestjs/swagger';
 import { Transform } from 'class-transformer';
 import { IsBoolean, IsOptional } from 'class-validator';
-import { PaginationQuery } from '@src/common';
+import { BasePaginationQuery } from '@src/common';
 
-export class GetPostQueryDto extends PaginationQuery {
+export class GetPostQueryDto extends BasePaginationQuery {
   @ApiProperty({
     required: false,
     description: '읽음 필터링 여부',
